text_extraction/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyJWT==2.8.0
motor==3.3.2
pymongo==4.7.2
openai>=1.14.0,<2.0.0
httpx>=0.27.0,<0.28.0
python-multipart==0.0.9
python-dotenv==1.0.1
PyYAML==6.0.1
passlib
redis==5.0.4
google-cloud-storage==2.10.0  # Cloud Storage for file uploads
google-cloud-pubsub==2.18.4   # Pub/Sub for message queuing
google-cloud-firestore==2.13.1  # Firestore for document storage
pdf2image==1.17.0  # PDF to image conversion
Pillow>=10.0.0  # Image processing
PyPDF2==3.0.1  # PDF metadata extraction